+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Refined iron recipe missing [#89]
+
 ## [0.7.3] - 2020-06-28
 Regenerated all tags and crushing/smelting recipes with data generators, tweaked some values
 ### Added